% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frappeChart.R
\name{frappeChart}
\alias{frappeChart}
\title{Create a Frappe Chart}
\usage{
frappeChart(
  data,
  type = c("line", "bar", "pie", "percentage", "heatmap"),
  title = "",
  ...,
  is_navigable = TRUE,
  width = NULL,
  height = 250,
  elementId = NULL
)
}
\description{
Creates a Frappe Chart
}
